% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accident_2015.R
\docType{data}
\name{accident_2015}
\alias{accident_2015}
\title{Fatality Analysis Database}
\format{An object of class tibble.}
\usage{
data(accident_2015)
}
\description{
data from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census providing the American public yearly data regarding fatal injuries suffered
in motor vehicle traffic crashes.
}
\keyword{datasets}
