% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state
The fars_map_state plots the accident locations of the years 2013, 2014,and 2015 per state.
The state is called by its number}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number as an integer, number must be between 1 and 51}

\item{year}{year as a string or an integer}
}
\value{
returns a plot of accident locations in the state called for the year called.
}
\description{
fars_map_state
The fars_map_state plots the accident locations of the years 2013, 2014,and 2015 per state.
The state is called by its number
}
\note{
Receives data from make_filename and fars_read functions. There are only 51 states in the USA. If a wrong state number is entered an error message will state "invalid STATE number: " with the wrong number printed.
If there is no data for a specific year, there is an error message saying "invalid year: " with the year printed.
If no accident occurred in the called state and year, the error message "no accidents to plot" appears.
Accidents for which the database has no data with regurds to LONGITUD and LATITUDE are placed outside of all potentital plots
}
\examples{
\dontrun{fars_map_state(51, 2013)} plots the accident locations for state 51 in year 2013.
}
