% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years
lists the month and the year for each accident of the year called}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The argument of the funtion is a vector of years}
}
\value{
returns a tibble (data.frame) of the length (number of accidents) x 2, whereas the columns return the month and the year in which the accident occured.
}
\description{
fars_read_years
lists the month and the year for each accident of the year called
}
\note{
Receives name of the file from the make_filename function, and selects months. If there is no data for a specific year, there is an error message saying "invalid year: " with the year printed.
}
\examples{
\dontrun{fars_read_years(2014)}
# [[1]]
# A tibble: 30,056 x 2
# MONTH  year
# <dbl> <dbl>
# 1     1  2014
# 2     1  2014
# 3     1  2014
# 4     1  2014
# 5     1  2014
# 6     1  2014
# 7     1  2014
# 8     1  2014
# 9     1  2014
# 10     1  2014
# ... with 30,046 more rows


}
