% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename
creates the file name for the called year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The argument of the funtion is a string or an integer with the year}
}
\value{
The function Prints sprintf("accident_%d.csv.bz2", year), hence returns a string
}
\description{
make_filename
creates the file name for the called year
}
\note{
the argument gets converted into an integer number and will then be used in the name which will be printed.
}
\examples{
\donttest{
 library(farsHK)
 make_filename(2013)
 } returns "accident_2013.csv.bz2"

}
